split vp direction out again, run june
